feat: Refactor balance deletion logic to ensure only the latest active balance can be deleted

+++ app/Services/BalanceService/BalanceDelete.php
@@ -13,16 +13,40 @@ class BalanceDelete
     {
         DB::beginTransaction();
         try {
-            $model = Balance::findOrFail($id);
-            $this->deleteImage($model->image);
 
-            // Delete agent and user
-            $model->delete();
+             $latestBalance = Balance::where('school_username', $school_username)->where('status', 1)
+             ->orderBy('updated_at', 'desc')
+             ->first();
+          $model = Balance::findOrFail($id);
+          if ($latestBalance && $latestBalance->id === $id) {
+               $this->deleteImage($model->image);
+               // Delete agent and user
+                $model->delete();
 
-            DB::commit();
+               DB::commit();
             return response()->json([
                 'message' => 'Balance deleted successfully',
             ], 200);
+
+           }else if($model->status == 0){
+                 $this->deleteImage($model->image);
+                // Delete agent and user
+                 $model->delete();
+
+                  DB::commit();
+                 return response()->json([
+                  'message' => 'Balance deleted successfully',
+               ], 200);
+           }
+           
+           else{
+               return response()->json([
+                    'message' => 'Only the latest active balance can be Deleted.',
+                    'latest_id' => $latestBalance ? $latestBalance->id : null,
+                ], 400);
+            }
+
+          
         } catch (\Exception $e) {
             DB::rollback();
             return response()->json([
